some minor bug fix

+++ src/main/java/com/zikozee/swaggerest/Repository/BallRepository.java
@@ -3,5 +3,5 @@ package com.zikozee.swaggerest.Repository;
 import com.zikozee.swaggerest.model.Ballers;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BallRepository extends CrudRepository<Ballers, Integer> {
+public interface BallRepository extends CrudRepository<Ballers, Long> {
 }
